@@ -15,81 +15,71 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.4+16
+version: 1.0.6+19
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0" # Updated to a more recent version for better compatibility
 
 dependencies:
   flutter:
     sdk: flutter
-  get: 4.6.1
-  global_configuration: 2.0.0-nullsafety.1
-  google_fonts: 3.0.1
-  velocity_x: 3.4.0
-  get_storage: 2.0.3
-  fluttertoast: 8.0.9
-  awesome_dialog: 2.2.1
-  change_app_package_name: 1.1.0
-  flutter_html: 3.0.0-alpha.3
-  cached_network_image: 3.2.0
-  flutter_easyloading: 3.0.3
-#  firebase_core_platform_interface: 4.5.1
-#  flutter_native_splash: 2.0.1+1
-#  mobile_number: ^2.1.1
-  flutter_svg: 1.0.3
-  pin_code_fields: 7.4.0
-  http: 0.13.4
-  carousel_slider: 4.1.1
-  awesome_bottom_bar:
-    path: awesome_bottom_bar
-  google_sign_in: 5.1.0
-  flutter_facebook_auth: 4.3.0
-  url_launcher: 6.1.5
-  flutter_barcode_scanner: 2.0.0
-  flutter_inappwebview: 5.4.0
-  upgrader: 4.7.0
-  assets_audio_player:
-    path: assets_audio_player-3.0.4+1
-  readmore: 2.1.0
-  youtube_player_iframe: 2.2.2
-#  modjs_video_player: ^1.0.0+fix
-  sms_autofill: 2.2.0
-  scrollable_positioned_list: 0.3.1
-  firebase_dynamic_links: 4.2.1
-  sign_in_with_apple: 3.3.0
-  dropdown_search: 3.0.1
-#  geolocator: 8.2.0
-  location: 4.4.0
-#  flutter_blue_plus: ^1.3.1
-  flutter_beacon:
-    path: flutter_beacon-0.5.1
-  flutter_speed_dial: 6.0.0
-  photo_view: 0.14.0
-  permission_handler: 9.2.0
-  cupertino_icons: 1.0.4
-  flutter_native_splash: 2.1.6
-  firebase_core_platform_interface: 4.5.1
+  get: ^4.6.1
+  global_configuration: ^2.0.0
+  google_fonts: 6.1.0
+  velocity_x: ^4.1.2
+  get_storage: ^2.0.3
+  fluttertoast: ^8.0.9
+  awesome_dialog: ^3.2.0
+  change_app_package_name: ^1.1.0
+  flutter_html: ^3.0.0-alpha.3
+  cached_network_image: ^3.3.1
+  flutter_easyloading: ^3.0.3
+  flutter_svg: ^2.0.10+1
+  pin_code_fields: ^7.4.0
+  http: ^1.2.1
+  carousel_slider: ^4.1.1
+  awesome_bottom_bar: any
+  google_sign_in: ^6.2.1
+  flutter_facebook_auth: ^4.3.0
+  url_launcher: ^6.1.0
+  flutter_barcode_scanner: ^2.0.0
+  flutter_inappwebview: ^5.4.0
+  upgrader: ^9.0.0
+  assets_audio_player: ^3.1.1
+  readmore: ^2.2.0
+  youtube_player_iframe: ^2.2.2
+  scrollable_positioned_list: ^0.3.1
+  firebase_dynamic_links: ^5.4.16
+  sign_in_with_apple: ^4.3.0
+  dropdown_search: ^3.0.1
+  location: ^5.0.3
+  flutter_beacon: ^0.5.1
+  flutter_speed_dial: ^6.0.0
+  photo_view: ^0.14.0
+  permission_handler: ^9.2.0
+  cupertino_icons: ^1.0.4
+  firebase_core_platform_interface: ^5.0.0
+  firebase_core: ^2.26.0
   flutter_cache_manager: ^3.3.0
-  flutter_downloader: 1.7.4
+  flutter_downloader: ^1.11.7
   path_provider: ^2.0.11
-  firebase_messaging: 11.2.15
-  flutter_local_notifications: 9.3.0
+  firebase_messaging: ^14.7.18
+  flutter_local_notifications: ^9.3.0
+  flutter_native_splash: ^2.3.10
+#  sms_autofill:
+#    git:
+#      url: https://github.com/jaumard/sms_autofill
+#      ref: master
+
+  flutter_callkit_incoming: ^2.0.4
+#  flutter_sms: ^2.3.3
+#  readsms: ^0.1.0+3
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
   assets:
     - assets/cfg/
     - assets/images/
