@@ -10,7 +10,7 @@ import 'package:url_launcher/url_launcher_string.dart';
 import '../core.dart';
 
 class CategoryTypesView extends GetView<CategoryTypeController> {
-  CategoryTypesView({Key key}) : super(key: key);
+  CategoryTypesView({Key? key}) : super(key: key);
   GlobalKey<ScaffoldState> scaffoldDashboardKey =
       GlobalKey<ScaffoldState>(debugLabel: '_scaffoldDashboardKey');
   CommonService commonService = Get.find();
@@ -45,7 +45,7 @@ class CategoryTypesView extends GetView<CategoryTypeController> {
             actions: [
               InkWell(
                 onTap: () {
-                  scaffoldDashboardKey.currentState.openEndDrawer();
+                  scaffoldDashboardKey.currentState?.openEndDrawer();
                 },
                 child: SvgPicture.asset(
                   "assets/images/menu.svg",
@@ -68,7 +68,7 @@ class CategoryTypesView extends GetView<CategoryTypeController> {
                 ),
                 commonService.selectedCategoryName.value.text
                     .textStyle(
-                      Get.textTheme.headline1.copyWith(
+                      Get.textTheme.headline1!.copyWith(
                         color: Get.theme.indicatorColor,
                         fontSize: 30,
                       ),
@@ -108,7 +108,7 @@ class CategoryTypesView extends GetView<CategoryTypeController> {
                                 height: 65,
                                 child: currentValue.title.text
                                     .textStyle(
-                                      Get.textTheme.headline2.copyWith(
+                                      Get.textTheme.headline2!.copyWith(
                                         color: Get.theme.indicatorColor,
                                       ),
                                     )
@@ -167,10 +167,10 @@ class CategoryTypesView extends GetView<CategoryTypeController> {
                                     height: controller
                                                 .firstVisibleItemIndex.value ==
                                             ind
-                                        ? 212
-                                        : 192,
+                                        ? 200
+                                        : 180,
                                     width: Get.width,
-                                    attributes: item,
+                                    attributes: item, showTitle: false, showVideoIcon: false,
                                   ).pOnly(bottom: 0),
                                 );
                               },
