@@ -3,14 +3,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
+import 'package:get_storage/get_storage.dart';
 import 'package:readmore/readmore.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
+import 'package:url_launcher/url_launcher_string.dart';
 import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
 import '../core.dart';
 
 class DetailView extends GetView<DetailController> {
-  DetailView({Key key}) : super(key: key);
+  DetailView({Key? key}) : super(key: key);
   GlobalKey<ScaffoldState> scaffoldDashboardKey = GlobalKey<ScaffoldState>(debugLabel: '_scaffoldDashboardKey');
   CommonService commonService = Get.find();
   @override
@@ -44,7 +46,7 @@ class DetailView extends GetView<DetailController> {
             actions: [
               InkWell(
                 onTap: () {
-                  scaffoldDashboardKey.currentState.openEndDrawer();
+                  scaffoldDashboardKey.currentState?.openEndDrawer();
                 },
                 child: SvgPicture.asset(
                   "assets/images/menu.svg",
@@ -89,7 +91,7 @@ class DetailView extends GetView<DetailController> {
                           ),
                           child: commonService.locationDetailData.value.attributes.locationName.text
                               .textStyle(
-                                Get.textTheme.headline1.copyWith(
+                                Get.textTheme.headline1!.copyWith(
                                   color: Get.theme.indicatorColor,
                                   fontSize: 30,
                                 ),
@@ -124,7 +126,7 @@ class DetailView extends GetView<DetailController> {
                                             : InkWell(
                                                 onTap: () {
                                                   AssetsAudioPlayer.allPlayers().forEach((key, value) {
-                                                  //  value.pause();
+                                                    value.pause();
                                                   });
                                                   commonService.assetsAudioPlayer.value.playOrPause();
                                                 },
@@ -155,15 +157,12 @@ class DetailView extends GetView<DetailController> {
                           ],
                         ).pSymmetric(h: 25)
                       : const SizedBox(height: 0),
-                  const SizedBox(
-                    height: 15,
-                  ),
                   SizedBox(
                     height: commonService.selectedAudioFile.value.isNotEmpty ? 20 : 0,
                   ),
                   commonService.locationDetailData.value.attributes.locationName.text
                       .textStyle(
-                        Get.textTheme.headline1.copyWith(
+                        Get.textTheme.headline1!.copyWith(
                           color: Get.theme.indicatorColor,
                           fontSize: 22,
                         ),
@@ -242,7 +241,7 @@ class DetailView extends GetView<DetailController> {
                                               fit: BoxFit.cover,
                                               width: 150,
                                               height: 150,
-                                              errorBuilder: (BuildContext context, Object exception, StackTrace stackTrace) {
+                                              errorBuilder: (BuildContext context, Object exception, StackTrace? stackTrace) {
                                                 return Image.asset(
                                                   Helper.localAssetPath(beaconItem.locationImage),
                                                   width: 150,
@@ -256,7 +255,7 @@ class DetailView extends GetView<DetailController> {
                                         width: 150,
                                         child: beaconItem.locationName.text
                                             .textStyle(
-                                              Get.textTheme.headline3.copyWith(
+                                              Get.textTheme.headline3!.copyWith(
                                                 color: Get.theme.indicatorColor,
                                               ),
                                             )
