@@ -11,7 +11,7 @@ import 'package:url_launcher/url_launcher_string.dart';
 import '../core.dart';
 
 class GuestInformationHistoryView extends GetView<GuestController> {
-  GuestInformationHistoryView({Key key}) : super(key: key);
+  GuestInformationHistoryView({Key? key}) : super(key: key);
   GlobalKey<ScaffoldState> scaffoldDashboardKey =
       GlobalKey<ScaffoldState>(debugLabel: '_scaffoldDashboardKey');
   CommonService commonService = Get.find();
@@ -56,7 +56,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
             actions: [
               InkWell(
                 onTap: () {
-                  scaffoldDashboardKey.currentState.openEndDrawer();
+                  scaffoldDashboardKey.currentState?.openEndDrawer();
                 },
                 child: SvgPicture.asset(
                   "assets/images/menu.svg",
@@ -89,7 +89,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                     controller.forUser.value
                         ? commonService.labelData.value.data.history.text.center
                             .textStyle(
-                              Get.textTheme.headline2.copyWith(
+                              Get.textTheme.headline2!.copyWith(
                                 color: Get.theme.indicatorColor,
                               ),
                             )
@@ -101,7 +101,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                               commonService.labelData.value.data.guestsEntered
                                   .text.center
                                   .textStyle(
-                                    Get.textTheme.headline2.copyWith(
+                                    Get.textTheme.headline2!.copyWith(
                                       color: Get.theme.indicatorColor,
                                     ),
                                   )
@@ -184,7 +184,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                           commonService.labelData.value.data
                                               .name.text.center
                                               .textStyle(
-                                                Get.textTheme.headline4
+                                                Get.textTheme.headline4!
                                                     .copyWith(
                                                   color: Get.theme.colorScheme
                                                       .primary,
@@ -196,7 +196,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                               .text
                                               .center
                                               .textStyle(
-                                                Get.textTheme.headline4
+                                                Get.textTheme.headline4!
                                                     .copyWith(
                                                   color: Get.theme.colorScheme
                                                       .primary,
@@ -206,7 +206,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                               .pOnly(right: 15),
                                           currentValue.name.text.center
                                               .textStyle(
-                                                Get.textTheme.bodyMedium
+                                                Get.textTheme.bodyMedium!
                                                     .copyWith(
                                                   color:
                                                       Get.theme.indicatorColor,
@@ -225,7 +225,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                           commonService.labelData.value.data
                                               .guests.text.center
                                               .textStyle(
-                                                Get.textTheme.headline4
+                                                Get.textTheme.headline4!
                                                     .copyWith(
                                                   color: Get.theme.colorScheme
                                                       .primary,
@@ -237,7 +237,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                               .text
                                               .center
                                               .textStyle(
-                                                Get.textTheme.headline4
+                                                Get.textTheme.headline4!
                                                     .copyWith(
                                                   color: Get.theme.colorScheme
                                                       .primary,
@@ -247,7 +247,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                               .pOnly(right: 15),
                                           currentValue.guests.text.center
                                               .textStyle(
-                                                Get.textTheme.bodyMedium
+                                                Get.textTheme.bodyMedium!
                                                     .copyWith(
                                                   color:
                                                       Get.theme.indicatorColor,
@@ -266,7 +266,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                             .text
                                             .center
                                             .textStyle(
-                                              Get.textTheme.bodyMedium.copyWith(
+                                              Get.textTheme.bodyMedium!.copyWith(
                                                 color: Get.theme.indicatorColor,
                                               ),
                                             )
@@ -295,10 +295,10 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                                         .theme
                                                         .colorScheme
                                                         .primary,
-                                                    context: Get.context,
-                                                    animType: AnimType.SCALE,
+                                                    context: Get.context!,
+                                                    animType: AnimType.scale,
                                                     dialogType:
-                                                        DialogType.QUESTION,
+                                                        DialogType.question,
                                                     body: Padding(
                                                       padding:
                                                           const EdgeInsets.only(
@@ -322,7 +322,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                                               .center
                                                               .textStyle(Get
                                                                   .textTheme
-                                                                  .headline1
+                                                                  .headline1!
                                                                   .copyWith(
                                                                       color: Get
                                                                           .theme
@@ -342,7 +342,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                                                               .center
                                                               .textStyle(Get
                                                                   .textTheme
-                                                                  .bodyText1
+                                                                  .bodyText1!
                                                                   .copyWith(
                                                                       color: Get
                                                                           .theme
@@ -460,7 +460,7 @@ class GuestInformationHistoryView extends GetView<GuestController> {
                         child: "No any history"
                             .text
                             .center
-                            .textStyle(Get.theme.textTheme.bodyMedium
+                            .textStyle(Get.theme.textTheme.bodyMedium!
                                 .copyWith(color: Get.theme.indicatorColor))
                             .make()
                             .centered(),
