@@ -1,25 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-// import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
-
+import '../controllers/bottom_nav_controller.dart';
 import '../core.dart';
 
 class LanguageView extends GetView {
-  LanguageView({Key key}) : super(key: key);
-  DateTime currentBackPressTime;
+  LanguageView({Key? key}) : super(key: key);
+  late DateTime currentBackPressTime;
   CategoryTypeController categoryTypeController = Get.find();
   DetailController detailController = Get.find();
   NearByController nearByController = Get.find();
   DashboardController dashboardController = Get.find();
   GuestController guestController = Get.find();
   CommonService commonService = Get.find();
+  BottomNavController controller= Get.put(BottomNavController());
+
   @override
   Widget build(BuildContext context) {
     SystemChrome.setSystemUIOverlayStyle(
-      SystemUiOverlayStyle(statusBarColor: Get.theme.primaryColor, statusBarIconBrightness: Brightness.dark),
+      SystemUiOverlayStyle(
+          statusBarColor: Get.theme.primaryColor,
+          statusBarIconBrightness: Brightness.dark),
     );
     return Scaffold(
       backgroundColor: Get.theme.primaryColorDark,
@@ -29,14 +32,21 @@ class LanguageView extends GetView {
         foregroundColor: Get.theme.indicatorColor,
         backgroundColor: Colors.transparent,
         elevation: 0,
-        automaticallyImplyLeading: (GetStorage().read('isBackToHomeLangPage') == '' || GetStorage().read('isBackToHomeLangPage') == null) ? false : true,
+        automaticallyImplyLeading:
+            (GetStorage().read('isBackToHomeLangPage') == '' ||
+                    GetStorage().read('isBackToHomeLangPage') == null)
+                ? false
+                : true,
         centerTitle: false,
       ),
       body: WillPopScope(
         onWillPop: () {
-          if (GetStorage().read('isBackToHomeLangPage') == '' || GetStorage().read('isBackToHomeLangPage') == null) {
+          if (GetStorage().read('isBackToHomeLangPage') == '' ||
+              GetStorage().read('isBackToHomeLangPage') == null) {
             DateTime now = DateTime.now();
-            if (currentBackPressTime == null || now.difference(currentBackPressTime) > const Duration(seconds: 2)) {
+            if (currentBackPressTime == null ||
+                now.difference(currentBackPressTime) >
+                    const Duration(seconds: 2)) {
               currentBackPressTime = now;
               Fluttertoast.showToast(msg: "Tap again to exit an app.");
               return Future.value(false);
@@ -65,7 +75,7 @@ class LanguageView extends GetView {
               ),
               commonService.labelData.value.data.selectLanguage.text
                   .textStyle(
-                    Get.textTheme.headline1.copyWith(
+                    Get.textTheme.headline1!.copyWith(
                       color: Get.theme.indicatorColor,
                       fontSize: 35,
                     ),
@@ -75,7 +85,7 @@ class LanguageView extends GetView {
                   .pOnly(bottom: 10),
               commonService.labelData.value.data.selectLangDesc.text
                   .textStyle(
-                    Get.textTheme.bodyText1.copyWith(
+                    Get.textTheme.bodyText1!.copyWith(
                       color: Get.theme.indicatorColor.withOpacity(0.5),
                       fontSize: 18,
                     ),
@@ -93,7 +103,7 @@ class LanguageView extends GetView {
                   categoryTypeController.fetchTrailsData();
                   detailController.getOfflineLocationsData();
                   nearByController.getNearByData();
-                  // commonService.firstTime=false;
+                  controller.updateIndex();
                   if (GetStorage().read('user_type') != "U") {
                     Get.toNamed('/emp_dashboard');
                     guestController.page = 1;
@@ -125,7 +135,7 @@ class LanguageView extends GetView {
                   child: "ਪੰਜਾਬੀ"
                       .text
                       .textStyle(
-                        Get.textTheme.headline3.copyWith(
+                        Get.textTheme.headline3!.copyWith(
                           color: Get.theme.highlightColor,
                         ),
                       )
@@ -141,7 +151,7 @@ class LanguageView extends GetView {
                   categoryTypeController.fetchTrailsData();
                   detailController.getOfflineLocationsData();
                   nearByController.getNearByData();
-                  // commonService.firstTime=false;
+                  controller.updateIndex();
 
                   if (GetStorage().read('user_type') != "U") {
                     Get.toNamed('/emp_dashboard');
@@ -174,7 +184,7 @@ class LanguageView extends GetView {
                   child: "हिन्दी"
                       .text
                       .textStyle(
-                        Get.textTheme.headline3.copyWith(
+                        Get.textTheme.headline3!.copyWith(
                           color: Get.theme.highlightColor,
                         ),
                       )
@@ -190,7 +200,7 @@ class LanguageView extends GetView {
                   categoryTypeController.fetchTrailsData();
                   detailController.getOfflineLocationsData();
                   nearByController.getNearByData();
-                  // commonService.firstTime=false;
+                  controller.updateIndex();
 
                   if (GetStorage().read('user_type') != "U") {
                     Get.toNamed('/emp_dashboard');
@@ -223,7 +233,7 @@ class LanguageView extends GetView {
                   child: "English"
                       .text
                       .textStyle(
-                        Get.textTheme.headline3.copyWith(
+                        Get.textTheme.headline3!.copyWith(
                           color: Get.theme.highlightColor,
                         ),
                       )
