@@ -1,4 +1,4 @@
 library globals;
 
 String eula =
-    '<p align="center"><strong>Bhagat Singh Museum App End User License Agreement (EULA)</strong><br>This End User License Agreement (&ldquo;Agreement&rdquo;) is between you and Bhagat Singh Museum and Punjab Heritage &amp; Tourism Promotion Board that solely governs the use of this Mobile and Web App made available through various App and online stores including but not limited to the Apple App Store and Google Play Store. By installing and or using the Bhagat Singh Museum App or web solution, you agree to be bound by this agreement.<br>1. Definitions<br>&ldquo;App Store&rdquo; means Google Play store or the Apple App Store.<br>&ldquo;App&rdquo; means Mobile or Web App.<br>2. Acknowledgement<br>a. The terms of this agreement Apply to the App or any of the services accessible through the App services, including any updates or supplements to the App or any service, unless they come with separate terms, in which case those terms Apply. If any open-source software is included in the App or any service, the terms of an open-source licence may override some of the terms of this agreement.<br>b. The new terms may be displayed on-screen and you may be required to read and accept them to continue your use of the services.<br>c. From time to time updates to the App may be issued through the App Store. Depending on the update, you may not be able to use the services until you have downloaded the latest version of the App and accepted any new terms.<br>d. You will be assumed to have obtained permission from the owners of the mobile telephone or handheld devices that are controlled, but not owned, by you (we refer to devices) and to download a copy of the App onto the devices. You and they may be charged by your and their service providers for internet access on the devices. You accept responsibility in accordance with the terms of this agreement for the use of the App or any service on or in relation to any device, whether or not it is owned by you.<br>e. By using the App or any of the services, you consent to us collecting and using technical information about the devices and related software, hardware and peripherals for services that are internet-based or wireless to improve our products and to provide any services to you.<br>f. By using the App or any of the services you also content to us collecting analytical data regarding the App or service usage.<br>g. The App or any service may contain links to other independent third-party websites (third-party sites such as flight check-In from airlines). Third-party sites are not under our control, and we are not responsible for and do not endorse their content or their privacy policies (if any). You will need to make your own independent judgement regarding your interaction with any third-party sites, including the purchase and use of any products or services accessible through them.<br>h. The App is developed and optimized for mobile and desktop, excluding tablet devices such as iPad.<br>i. If you are a user assigned to a corporate account, you acknowledge that your subscription level will be set back to &ldquo;Free&rdquo; in case the corporate account gets suspended or the corporate account admin removes you from the account.<br>3. Parties<br>This Agreement is between you and Bhagat Singh Museum only. Notwithstanding the foregoing, you acknowledge that Apple, Google, and their respective subsidiaries are third party beneficiaries of this agreement and the forenamed parties have the right to enforce this agreement against you.<br>4. Age Restrictions<br>By using the Bhagat Singh Museum App, you represent and warrant that<br>(a) you are 17 years of age or older and you agree to be bound by this Agreement;<br>(b) if you are under 17 years of age, you have obtained verifiable consent from a parent or legal guardian; and<br>(c) your use of the Bhagat Singh Museum App does not violate any Applicable law or regulation. Your access to the Bhagat Singh Museum App may be terminated without warning if Bhagat Singh Museum has plausible reasons to believe that you are under the age of 17 years and have not obtained verifiable consent from a parent or legal guardian. If you are a parent or legal guardian and you provide your consent to your child&rsquo;s use of the Bhagat Singh Museum App, you agree to be bound by this agreement in respect to your child&rsquo;s use of the Bhagat Singh Museum App.<br>5. Objectionable Content Policy<br>Content submitted to Bhagat Singh Museum may be moderated. Content which directly or indirectly involves any objectionable Content will be immediately removed. Objectionable content includes, but is not limited to: (i) sexually explicit materials; (ii) obscene, defamatory, libelous, slanderous, violent and/or unlawful content or profanity; (iii) content that infringes upon the rights of any third party, including copyright, trademark, privacy, publicity or other personal or proprietary right, or that is deceptive or fraudulent; (iv) content that promotes the use or sale of illegal or regulated substances, tobacco products, ammunition and/or firearms; and (v) gambling, including without limitation, any online casino, sports books, bingo or poker.<br>6. Licence restrictions<br>Except as expressly set out in this agreement or as permitted by any local law, you agree:<br>i. not to copy the App or documents except where such copying is incidental to normal use of the App, or where it is necessary for the purpose of back-up or operational security;<br>ii. not to rent, lease, sub-license, loan, translate, merge, adapt, vary or modify the App;<br>iii. not to make alterations to, or modifications of, the whole or any part of the App, or permit the App or any part of it to be combined with, or become incorporated in, any other programs;<br>iv. not to disassemble, decompile, reverse-engineer or create derivative works based on the whole or any part of the App or attempt to do any such thing except to the extent that such actions cannot be prohibited because they are essential for the purpose of achieving inter-operability of the App with another software program, and provided that the information obtained by you during such activities:<br>a) is used only for the purpose of achieving inter-operability of the App with another software program;<br>b) is not unnecessarily disclosed or communicated without our prior written consent to any third party; and<br>c) is not used to create any software that is substantially similar to the App;<br>7. Acceptable use restrictions<br>You must:<br>i. not use the App or any service in any unlawful manner, for any unlawful purpose, or in any manner inconsistent with this agreement, or act fraudulently or maliciously, for example, by hacking into or inserting malicious code, including viruses, or harmful data, into the App, any service or any operating system;<br>ii. not infringe our intellectual property rights or those of any third party in relation to your use of the App or any service (to the extent that such use is not licensed by this agreement);<br>iii. not transmit any material that is defamatory, offensive or otherwise objectionable in relation to your use of the App or any service;<br>iv. not use the App or any service in a way that could damage, disable, overburden, impair or compromise our systems or security or interfere with other users; and<br>v. not collect or harvest any information or data from any service or our systems or attempt to decipher any transmissions to or from the servers running any service.<br>8. Intellectual property rights<br>a. You acknowledge that all intellectual property rights in the App and the technology anywhere in the world belong to us (Bhagat Singh Museum or Punjab Heritage &amp; Tourism Promotion Board) or our licensors, that rights in the App are licensed (not sold) to you, and that you have no rights in, or to, the App or the technology other than the right to use each of them in accordance with the terms of this agreement.<br>b. You acknowledge that you have no right to have access to the App in source-code form.<br>9. No warranty or support<br>a. You expressly acknowledge and agree that use of the App and any documents and services provided is at your sole risk and that the entire risk as to satisfactory quality, performance, accuracy and effort is with you.<br>b. To the maximum extent permitted by Applicable law, the App and documents and services are provided &quot;as is&quot; and &ldquo;as available&rdquo;, with all faults and without warranty of any kind, and we hereby disclaim all warranties and conditions with respect to the App, documents and services, either express, implied or statutory, including, but not limited to, any implied warranties and/or conditions of merchantability, satisfactory quality, fitness for a particular purpose, of accuracy, of quiet enjoyment, and non-infringement of third party rights. We do not warrant against interference with your enjoyment of the App or documents, that the functions contained in the App or services will meet your requirements, that the operation of the App or services will be uninterrupted or error-free, or that defects in the App or services will be corrected. No oral or written information or advice<br>given by us or our authorized representative shall create a warranty. Should the App or services prove defective, you assume the entire cost of all necessary servicing, repair or correction.<br>10. Limitation of liability<br>a. You acknowledge the App has not been developed to meet your requirements, and that it is, therefore, your full responsibility to ensure that the facilities and functions of the App as described in the documents meet your personal requirements.<br>b. The Bhagat Singh Museum App relies on internal and third-party services to provide you -<br>if any - with:<br>● Trip information<br>● Start and end date &amp; time of trip items<br>● Trail/Destination information<br>● Real-time alert and notification specific to Spot/Location<br>● Country specific travel risk advisories (including but not exclusive security risk,<br>medical risk, transport and weather risk)<br>● Destination address with link to map<br>c. Information listed above is non-exhaustive and can change at any time at the sole discretion of Bhagat Singh Museum. We are unable to guarantee you the accuracy, completeness, and timeliness of any information, alerts, and notifications provided or not provided to you. As a user, you are responsible for checking, controlling and verifying the accuracy, completeness, and timeliness of the information, alerts, and notifications provided or not provided to you via the official channel providers of the information in question (corporate and government sources). Bhagat Singh Museum is under no circumstances reliable for personal injury, illness, health problems, direct, indirect or consequential whatsoever, including without limitation, damages for loss of profits, loss of data, business interruption, costs or any other commercial damages or losses, arising out of or related to your use or inability to use our Application and services.<br>11 Termination<br>a) We may terminate this agreement immediately by written notice to you:<br>i) if you commit a material or persistent breach of this agreement which you fail to remedy (if remediable) within 14 days after the service of written notice requiring you to do so;<br>ii) if you breach any of the licence restrictions or the acceptable use restrictions; and<br>iii) on thirty days&rsquo; notice.<br>b) On termination for any reason:<br>i) all rights granted to you under this agreement shall cease;<br>ii) you must immediately cease all activities authorised by this agreement;<br>iii) you must immediately delete or remove the App from all devices, and immediately destroy all copies of the App and documents then in your possession, custody or control and certify to us that you have done so.<br>12 Products and Services from Business Partners, Affiliates and Third-Party Provider<br>Bhagat Singh Museum is not a travel agency, insurance agency, retailer, manufacturer, or service provider. Any services displayed are provided by business partners and/or their affiliates and/or their third-party providers.<br>You acknowledge that you are using our App at your own risk. We are not responsible for any content and accuracy of any third-party product and/or service, nor for advertisement, nor for information, nor for providing any travel options, displayed in our Applications or derivatives and/or in any business partner and/or their affiliates and/or their third-party provider sites.<br>We assume no liability in respect of such as, but not limited to, advertisement, information, offering, quote, discount, deal, booking, contract, confirmation, cancellation, policy and Application (If any of the mentioned exists).<br>Bhagat Singh Museum might share trip data such as, but not limited to, trip dates, destination, as well as email address and if needed for offering, nationality and residency, to third business partners and/or their affiliates and/or their third-party provider. That information might be shared only after you click/tap on a digital advertising instrument such as, but not limited to, pop-up, banner, link and call-to-action-box, displayed anywhere in the mobile and/or web application and/or digital communication. By clicking/tapping you irrevocably agree to a potential data transfer.<br>Business partners and/or their affiliates and/or their third-party providers, might use cookies to ensure that Bhagat Singh Museum might get a commission after product and/or service has been paid (If any of the mentioned exists) and/or used by the user. Every business partner and/or their affiliate and/or their third-party provider has its own privacy policy, end-user-license-agreement, terms &amp; conditions. We strongly encourage you to visit their website to read their privacy policy, end-user-license-agreement and terms &amp; conditions to find out how that Site/App collects and uses your information and to evaluate whether and for what purpose they use cookies.<br>13 Communication between us<br>If you wish to contact us in writing, or if any condition in this agreement requires you to give us notice in writing, you can send this to us by e-mail at<a href="mailto:info@dwgroup.in">info@dwgroup.in</a><br>If we have to contact you or give you notice in writing, we will do so by e-mail to the email address you provided to us in your request for the App.<br>14 Events outside our control<br>We will not be liable or responsible for any failure to perform, or delay in performance of, any of our obligations under this agreement that is caused by any act or event beyond our reasonable control, including failure of public or private telecommunications networks (event outside our control).<br>If an event outside our control takes place that affects the performance of our obligations under this Agreement:<br>(a) our obligations under this agreement will be suspended and the time for performance of our obligations will be extended for the duration of the event outside our control; and (b) we will use our reasonable endeavours to find a solution by which our obligations under this agreement may be performed despite the event outside our control.<br>15 Other important terms<br>We may transfer our rights and obligations under this agreement to another organisation (If any of the mentioned exists), but this will not affect your rights or our obligations under this agreement.<br>You may only transfer your rights or obligations under this agreement to another person if we agree in writing.<br>If we fail to insist that you perform any of your obligations under this agreement, or if we do not enforce our rights against you, or if we delay in doing so, that will not mean that we have waived our rights against you and will not mean that you do not have to comply with those obligations. If we do waive a default by you, we will only do so in writing, and that will not mean that we will automatically waive any later default by you.<br>Each of the conditions of this agreement operates separately. If any court or competent authority decides that any of them are unlawful or unenforceable, the remaining conditions will remain in full force and effect.<br>16 Updates and Validity<br>We may update this End - User License Agreement (EULA) from time to time to reflect the new features, products and services (If any of the mentioned exists) recently added in the Application and digital communication as well changes in Applicable laws and regulations. If we update our agreement, we increase the agreement version and update the release date.<br>For any version of the Application that you have downloaded before a EULA update, the latest agreement will be legally binding and an irrevocable requirement for the further use of the Application.<br>If you do not agree to the terms of this agreement available at Rupnagar.nic.in/tourist-places/virasta-e-khalsa/EULA as<br>updated from time to time, Bhagat Singh Museum is unwilling to license the App to you and you must<br>stop using our Applications, products and services and/or close the account. Your continued use<br>of our Applications, features, products and services will be deemed as your acceptance of any<br>agreement updates.<br>For any questions, kindly send a message to info.<br>17 Notifications<br>You consent and agree to receive notice of agreement updates through our posting for updated EULA at Rupnagar.nic.in/tourist-places/virasta-e-khalsa /EULA. We encourage you to periodically visit<br>Rupnagar.nic.in/tourist-places/virasta-e-khalsa /EULA for the latest information on our End User License Agreement.<br>We also might notify you by email (sent to the e-mail address specified in your account) and/or<br>within the App and/or digital communication instruments.<br>18 Translations<br>You agree that the English version of the Application including all content and the EULA will govern in the event of a conflict between it and any version translated into another language.<br>English Version:<a href="https://www.Bhagat Singh Museum.com/EULA/Bhagat Singh Museum-EULA.pdf">https://www.Bhagat Singh Museum.com/EULA/Bhagat Singh Museum-EULA.pdf</a><br>Important notice:<br>The Bhagat Singh Museum App stores and processes personal data that you have provided to us so that we can create and present trip-related information for your convenience. It&rsquo;s your responsibility to keep your phone and access to the App secure. We therefore recommend that you do not jailbreak or root your phone, which is the process of removing software restrictions and limitations imposed by the official operating system of your device. It could make your phone vulnerable to  malware/viruses/malicious programs, compromise your phone&rsquo;s security features and it could mean that the Bhagat Singh Museum App won&rsquo;t work properly or at all.<br>You should be aware that there are certain things that Bhagat Singh Museum will not take responsibility for. The App requires an active internet connection for certain functionalities. The connection can be Wi-Fi, or provided by your mobile network provider. Bhagat Singh Museum cannot take responsibility for the App not working at full functionality if you don&rsquo;t have access to Wi-Fi/Internet.<br>If you&rsquo;re using the App outside of an area with Wi-Fi, you should remember that your terms of agreement with your mobile network provider will still Apply. As a result, you may be charged by your mobile provider for the cost of data for the duration of the connection while accessing the App, or other third-party charges. In using the App, you&rsquo;re accepting responsibility for any such charges, including roaming data charges if you use the App outside of your home territory (i.e. region or country) without turning off data roaming. If you are not the bill payer for the device on which you&rsquo;re using the App, please be aware that we assume that you have received permission from the bill payer (if any of them exists) for using the App.<br>To ensure best possible use of the App, we strongly recommend to be up to date with the latest operating system and browser version.<br>Along the same lines, Bhagat Singh Museum cannot take responsibility for the way you use the App. You also need to make sure that your device stays charged – if it runs out of battery and you can&rsquo;t turn it on to travel updates and notifications, Bhagat Singh Museum cannot accept responsibility if you are as a result not able to travel.<br>You should also remember that the App&rsquo;s passport scanner (when made available) function&rsquo;s purpose is to make things more convenient for you – it doesn&rsquo;t replace your travel documents, and you will still need to bring them to the airport in order to fly.<br>Lastly, with respect to our responsibility for your use of the App, it&rsquo;s important to bear in mind that although we endeavour to ensure that it is updated and correct at all times, we do rely on third parties to provide information to us so as to provide you with useful updates when there are likely or confirmed changes. So, it is not intended to wholly replace the need to stay informed of updates that are made on screens throughout the terminal or via airport loudspeaker announcements. Therefore, we cannot accept liability for any loss, direct or indirect, you experience as a result of relying wholly on this functionality of the App rather than using all the resources in the airport to keep yourself updated as to the details of your flight.<br>By downloading the App from this site or clicking on the &quot;Accept&quot; button below you agree to the terms of the licence, stated above, which will bind you. Furthermore, this also includes your acceptance of our privacy policy as available on our website.<br>If you do not agree to the terms of this licence, we will not license the App to you and you must stop the downloading or streaming process (as Applicable) now.<br>EULA 1.1. 2022</p>';
+    '<p align="center"><strong>Dastaan-E-Shahadat App End User License Agreement (EULA)</strong><br>This End User License Agreement (&ldquo;Agreement&rdquo;) is between you and Virasat-E-Khalsa and Punjab Heritage &amp; Tourism Promotion Board that solely governs the use of this Mobile and Web App made available through various App and online stores including but not limited to the Apple App Store and Google Play Store. By installing and or using the Virasat-E-Khalsa App or web solution, you agree to be bound by this agreement.<br>1. Definitions<br>&ldquo;App Store&rdquo; means Google Play store or the Apple App Store.<br>&ldquo;App&rdquo; means Mobile or Web App.<br>2. Acknowledgement<br>a. The terms of this agreement Apply to the App or any of the services accessible through the App services, including any updates or supplements to the App or any service, unless they come with separate terms, in which case those terms Apply. If any open-source software is included in the App or any service, the terms of an open-source licence may override some of the terms of this agreement.<br>b. The new terms may be displayed on-screen and you may be required to read and accept them to continue your use of the services.<br>c. From time to time updates to the App may be issued through the App Store. Depending on the update, you may not be able to use the services until you have downloaded the latest version of the App and accepted any new terms.<br>d. You will be assumed to have obtained permission from the owners of the mobile telephone or handheld devices that are controlled, but not owned, by you (we refer to devices) and to download a copy of the App onto the devices. You and they may be charged by your and their service providers for internet access on the devices. You accept responsibility in accordance with the terms of this agreement for the use of the App or any service on or in relation to any device, whether or not it is owned by you.<br>e. By using the App or any of the services, you consent to us collecting and using technical information about the devices and related software, hardware and peripherals for services that are internet-based or wireless to improve our products and to provide any services to you.<br>f. By using the App or any of the services you also content to us collecting analytical data regarding the App or service usage.<br>g. The App or any service may contain links to other independent third-party websites (third-party sites such as flight check-In from airlines). Third-party sites are not under our control, and we are not responsible for and do not endorse their content or their privacy policies (if any). You will need to make your own independent judgement regarding your interaction with any third-party sites, including the purchase and use of any products or services accessible through them.<br>h. The App is developed and optimized for mobile and desktop, excluding tablet devices such as iPad.<br>i. If you are a user assigned to a corporate account, you acknowledge that your subscription level will be set back to &ldquo;Free&rdquo; in case the corporate account gets suspended or the corporate account admin removes you from the account.<br>3. Parties<br>This Agreement is between you and Virasat-E-Khalsa only. Notwithstanding the foregoing, you acknowledge that Apple, Google, and their respective subsidiaries are third party beneficiaries of this agreement and the forenamed parties have the right to enforce this agreement against you.<br>4. Age Restrictions<br>By using the Virasat-E-Khalsa App, you represent and warrant that<br>(a) you are 17 years of age or older and you agree to be bound by this Agreement;<br>(b) if you are under 17 years of age, you have obtained verifiable consent from a parent or legal guardian; and<br>(c) your use of the Virasat-E-Khalsa App does not violate any Applicable law or regulation. Your access to the Virasat-E-Khalsa App may be terminated without warning if Virasat-E-Khalsa has plausible reasons to believe that you are under the age of 17 years and have not obtained verifiable consent from a parent or legal guardian. If you are a parent or legal guardian and you provide your consent to your child&rsquo;s use of the Virasat-E-Khalsa App, you agree to be bound by this agreement in respect to your child&rsquo;s use of the Virasat-E-Khalsa App.<br>5. Objectionable Content Policy<br>Content submitted to Virasat-E-Khalsa may be moderated. Content which directly or indirectly involves any objectionable Content will be immediately removed. Objectionable content includes, but is not limited to: (i) sexually explicit materials; (ii) obscene, defamatory, libelous, slanderous, violent and/or unlawful content or profanity; (iii) content that infringes upon the rights of any third party, including copyright, trademark, privacy, publicity or other personal or proprietary right, or that is deceptive or fraudulent; (iv) content that promotes the use or sale of illegal or regulated substances, tobacco products, ammunition and/or firearms; and (v) gambling, including without limitation, any online casino, sports books, bingo or poker.<br>6. Licence restrictions<br>Except as expressly set out in this agreement or as permitted by any local law, you agree:<br>i. not to copy the App or documents except where such copying is incidental to normal use of the App, or where it is necessary for the purpose of back-up or operational security;<br>ii. not to rent, lease, sub-license, loan, translate, merge, adapt, vary or modify the App;<br>iii. not to make alterations to, or modifications of, the whole or any part of the App, or permit the App or any part of it to be combined with, or become incorporated in, any other programs;<br>iv. not to disassemble, decompile, reverse-engineer or create derivative works based on the whole or any part of the App or attempt to do any such thing except to the extent that such actions cannot be prohibited because they are essential for the purpose of achieving inter-operability of the App with another software program, and provided that the information obtained by you during such activities:<br>a) is used only for the purpose of achieving inter-operability of the App with another software program;<br>b) is not unnecessarily disclosed or communicated without our prior written consent to any third party; and<br>c) is not used to create any software that is substantially similar to the App;<br>7. Acceptable use restrictions<br>You must:<br>i. not use the App or any service in any unlawful manner, for any unlawful purpose, or in any manner inconsistent with this agreement, or act fraudulently or maliciously, for example, by hacking into or inserting malicious code, including viruses, or harmful data, into the App, any service or any operating system;<br>ii. not infringe our intellectual property rights or those of any third party in relation to your use of the App or any service (to the extent that such use is not licensed by this agreement);<br>iii. not transmit any material that is defamatory, offensive or otherwise objectionable in relation to your use of the App or any service;<br>iv. not use the App or any service in a way that could damage, disable, overburden, impair or compromise our systems or security or interfere with other users; and<br>v. not collect or harvest any information or data from any service or our systems or attempt to decipher any transmissions to or from the servers running any service.<br>8. Intellectual property rights<br>a. You acknowledge that all intellectual property rights in the App and the technology anywhere in the world belong to us (Virasat-E-Khalsa or Punjab Heritage &amp; Tourism Promotion Board) or our licensors, that rights in the App are licensed (not sold) to you, and that you have no rights in, or to, the App or the technology other than the right to use each of them in accordance with the terms of this agreement.<br>b. You acknowledge that you have no right to have access to the App in source-code form.<br>9. No warranty or support<br>a. You expressly acknowledge and agree that use of the App and any documents and services provided is at your sole risk and that the entire risk as to satisfactory quality, performance, accuracy and effort is with you.<br>b. To the maximum extent permitted by Applicable law, the App and documents and services are provided &quot;as is&quot; and &ldquo;as available&rdquo;, with all faults and without warranty of any kind, and we hereby disclaim all warranties and conditions with respect to the App, documents and services, either express, implied or statutory, including, but not limited to, any implied warranties and/or conditions of merchantability, satisfactory quality, fitness for a particular purpose, of accuracy, of quiet enjoyment, and non-infringement of third party rights. We do not warrant against interference with your enjoyment of the App or documents, that the functions contained in the App or services will meet your requirements, that the operation of the App or services will be uninterrupted or error-free, or that defects in the App or services will be corrected. No oral or written information or advice<br>given by us or our authorized representative shall create a warranty. Should the App or services prove defective, you assume the entire cost of all necessary servicing, repair or correction.<br>10. Limitation of liability<br>a. You acknowledge the App has not been developed to meet your requirements, and that it is, therefore, your full responsibility to ensure that the facilities and functions of the App as described in the documents meet your personal requirements.<br>b. The Virasat-E-Khalsa App relies on internal and third-party services to provide you -<br>if any - with:<br>● Trip information<br>● Start and end date &amp; time of trip items<br>● Trail/Destination information<br>● Real-time alert and notification specific to Spot/Location<br>● Country specific travel risk advisories (including but not exclusive security risk,<br>medical risk, transport and weather risk)<br>● Destination address with link to map<br>c. Information listed above is non-exhaustive and can change at any time at the sole discretion of Virasat-E-Khalsa. We are unable to guarantee you the accuracy, completeness, and timeliness of any information, alerts, and notifications provided or not provided to you. As a user, you are responsible for checking, controlling and verifying the accuracy, completeness, and timeliness of the information, alerts, and notifications provided or not provided to you via the official channel providers of the information in question (corporate and government sources). Virasat-E-Khalsa is under no circumstances reliable for personal injury, illness, health problems, direct, indirect or consequential whatsoever, including without limitation, damages for loss of profits, loss of data, business interruption, costs or any other commercial damages or losses, arising out of or related to your use or inability to use our Application and services.<br>11 Termination<br>a) We may terminate this agreement immediately by written notice to you:<br>i) if you commit a material or persistent breach of this agreement which you fail to remedy (if remediable) within 14 days after the service of written notice requiring you to do so;<br>ii) if you breach any of the licence restrictions or the acceptable use restrictions; and<br>iii) on thirty days&rsquo; notice.<br>b) On termination for any reason:<br>i) all rights granted to you under this agreement shall cease;<br>ii) you must immediately cease all activities authorised by this agreement;<br>iii) you must immediately delete or remove the App from all devices, and immediately destroy all copies of the App and documents then in your possession, custody or control and certify to us that you have done so.<br>12 Products and Services from Business Partners, Affiliates and Third-Party Provider<br>Virasat-E-Khalsa is not a travel agency, insurance agency, retailer, manufacturer, or service provider. Any services displayed are provided by business partners and/or their affiliates and/or their third-party providers.<br>You acknowledge that you are using our App at your own risk. We are not responsible for any content and accuracy of any third-party product and/or service, nor for advertisement, nor for information, nor for providing any travel options, displayed in our Applications or derivatives and/or in any business partner and/or their affiliates and/or their third-party provider sites.<br>We assume no liability in respect of such as, but not limited to, advertisement, information, offering, quote, discount, deal, booking, contract, confirmation, cancellation, policy and Application (If any of the mentioned exists).<br>Virasat-E-Khalsa might share trip data such as, but not limited to, trip dates, destination, as well as email address and if needed for offering, nationality and residency, to third business partners and/or their affiliates and/or their third-party provider. That information might be shared only after you click/tap on a digital advertising instrument such as, but not limited to, pop-up, banner, link and call-to-action-box, displayed anywhere in the mobile and/or web application and/or digital communication. By clicking/tapping you irrevocably agree to a potential data transfer.<br>Business partners and/or their affiliates and/or their third-party providers, might use cookies to ensure that Virasat-E-Khalsa might get a commission after product and/or service has been paid (If any of the mentioned exists) and/or used by the user. Every business partner and/or their affiliate and/or their third-party provider has its own privacy policy, end-user-license-agreement, terms &amp; conditions. We strongly encourage you to visit their website to read their privacy policy, end-user-license-agreement and terms &amp; conditions to find out how that Site/App collects and uses your information and to evaluate whether and for what purpose they use cookies.<br>13 Communication between us<br>If you wish to contact us in writing, or if any condition in this agreement requires you to give us notice in writing, you can send this to us by e-mail at<a href="mailto:info@dwgroup.in">info@dwgroup.in</a><br>If we have to contact you or give you notice in writing, we will do so by e-mail to the email address you provided to us in your request for the App.<br>14 Events outside our control<br>We will not be liable or responsible for any failure to perform, or delay in performance of, any of our obligations under this agreement that is caused by any act or event beyond our reasonable control, including failure of public or private telecommunications networks (event outside our control).<br>If an event outside our control takes place that affects the performance of our obligations under this Agreement:<br>(a) our obligations under this agreement will be suspended and the time for performance of our obligations will be extended for the duration of the event outside our control; and (b) we will use our reasonable endeavours to find a solution by which our obligations under this agreement may be performed despite the event outside our control.<br>15 Other important terms<br>We may transfer our rights and obligations under this agreement to another organisation (If any of the mentioned exists), but this will not affect your rights or our obligations under this agreement.<br>You may only transfer your rights or obligations under this agreement to another person if we agree in writing.<br>If we fail to insist that you perform any of your obligations under this agreement, or if we do not enforce our rights against you, or if we delay in doing so, that will not mean that we have waived our rights against you and will not mean that you do not have to comply with those obligations. If we do waive a default by you, we will only do so in writing, and that will not mean that we will automatically waive any later default by you.<br>Each of the conditions of this agreement operates separately. If any court or competent authority decides that any of them are unlawful or unenforceable, the remaining conditions will remain in full force and effect.<br>16 Updates and Validity<br>We may update this End - User License Agreement (EULA) from time to time to reflect the new features, products and services (If any of the mentioned exists) recently added in the Application and digital communication as well changes in Applicable laws and regulations. If we update our agreement, we increase the agreement version and update the release date.<br>For any version of the Application that you have downloaded before a EULA update, the latest agreement will be legally binding and an irrevocable requirement for the further use of the Application.<br>If you do not agree to the terms of this agreement available at Rupnagar.nic.in/tourist-places/virasta-e-khalsa/EULA as<br>updated from time to time, Virasat-E-Khalsa is unwilling to license the App to you and you must<br>stop using our Applications, products and services and/or close the account. Your continued use<br>of our Applications, features, products and services will be deemed as your acceptance of any<br>agreement updates.<br>For any questions, kindly send a message to info.<br>17 Notifications<br>You consent and agree to receive notice of agreement updates through our posting for updated EULA at Rupnagar.nic.in/tourist-places/virasta-e-khalsa /EULA. We encourage you to periodically visit<br>Rupnagar.nic.in/tourist-places/virasta-e-khalsa /EULA for the latest information on our End User License Agreement.<br>We also might notify you by email (sent to the e-mail address specified in your account) and/or<br>within the App and/or digital communication instruments.<br>18 Translations<br>You agree that the English version of the Application including all content and the EULA will govern in the event of a conflict between it and any version translated into another language.<br>English Version:<a href="https://www.Virasat-E-Khalsa.com/EULA/Virasat-E-Khalsa-EULA.pdf">https://www.Virasat-E-Khalsa.com/EULA/Virasat-E-Khalsa-EULA.pdf</a><br>Important notice:<br>The Virasat-E-Khalsa App stores and processes personal data that you have provided to us so that we can create and present trip-related information for your convenience. It&rsquo;s your responsibility to keep your phone and access to the App secure. We therefore recommend that you do not jailbreak or root your phone, which is the process of removing software restrictions and limitations imposed by the official operating system of your device. It could make your phone vulnerable to  malware/viruses/malicious programs, compromise your phone&rsquo;s security features and it could mean that the Virasat-E-Khalsa App won&rsquo;t work properly or at all.<br>You should be aware that there are certain things that Virasat-E-Khalsa will not take responsibility for. The App requires an active internet connection for certain functionalities. The connection can be Wi-Fi, or provided by your mobile network provider. Virasat-E-Khalsa cannot take responsibility for the App not working at full functionality if you don&rsquo;t have access to Wi-Fi/Internet.<br>If you&rsquo;re using the App outside of an area with Wi-Fi, you should remember that your terms of agreement with your mobile network provider will still Apply. As a result, you may be charged by your mobile provider for the cost of data for the duration of the connection while accessing the App, or other third-party charges. In using the App, you&rsquo;re accepting responsibility for any such charges, including roaming data charges if you use the App outside of your home territory (i.e. region or country) without turning off data roaming. If you are not the bill payer for the device on which you&rsquo;re using the App, please be aware that we assume that you have received permission from the bill payer (if any of them exists) for using the App.<br>To ensure best possible use of the App, we strongly recommend to be up to date with the latest operating system and browser version.<br>Along the same lines, Virasat-E-Khalsa cannot take responsibility for the way you use the App. You also need to make sure that your device stays charged – if it runs out of battery and you can&rsquo;t turn it on to travel updates and notifications, Virasat-E-Khalsa cannot accept responsibility if you are as a result not able to travel.<br>You should also remember that the App&rsquo;s passport scanner (when made available) function&rsquo;s purpose is to make things more convenient for you – it doesn&rsquo;t replace your travel documents, and you will still need to bring them to the airport in order to fly.<br>Lastly, with respect to our responsibility for your use of the App, it&rsquo;s important to bear in mind that although we endeavour to ensure that it is updated and correct at all times, we do rely on third parties to provide information to us so as to provide you with useful updates when there are likely or confirmed changes. So, it is not intended to wholly replace the need to stay informed of updates that are made on screens throughout the terminal or via airport loudspeaker announcements. Therefore, we cannot accept liability for any loss, direct or indirect, you experience as a result of relying wholly on this functionality of the App rather than using all the resources in the airport to keep yourself updated as to the details of your flight.<br>By downloading the App from this site or clicking on the &quot;Accept&quot; button below you agree to the terms of the licence, stated above, which will bind you. Furthermore, this also includes your acceptance of our privacy policy as available on our website.<br>If you do not agree to the terms of this licence, we will not license the App to you and you must stop the downloading or streaming process (as Applicable) now.<br>EULA 1.1. 2022</p>';
